+++ model/Request.cs
@@ -1,0 +1,9 @@
+namespace AwsDotnetCsharp
+{
+    public class Request
+    {
+        public string Key1 { get; set; }
+        public string Key2 { get; set; }
+        public string Key3 { get; set; }
+    }
+}
